+++ rs/libmoq/src/consume.rs
@@ -206,6 +206,12 @@ impl Consume {
 			coded_width: config.coded_width.unwrap_or(0),
 			coded_height: config.coded_height.unwrap_or(0),
 			container: crate::api::borrow_container(&config.container),
+			label: config
+				.label
+				.as_ref()
+				.map(|label| label.as_ptr() as *const c_char)
+				.unwrap_or(std::ptr::null()),
+			label_len: config.label.as_ref().map_or(0, String::len),
 		};
 
 		Ok(())
@@ -268,6 +274,12 @@ impl Consume {
 			sample_rate: config.sample_rate,
 			channel_count: config.channel_count,
 			container: crate::api::borrow_container(&config.container),
+			label: config
+				.label
+				.as_ref()
+				.map(|label| label.as_ptr() as *const c_char)
+				.unwrap_or(std::ptr::null()),
+			label_len: config.label.as_ref().map_or(0, String::len),
 		};
 
 		Ok(())